% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_as_sleep_logs.R
\name{as_sleep_logs}
\alias{as_sleep_logs}
\title{Prepares raw data for later processing}
\usage{
as_sleep_logs(all_sleep_dat)
}
\arguments{
\item{all_sleep_dat}{sleep-levels dataset. Only the person_id and start_datetime are needed, but the full dataset is usually passed in containing \cr
person_id, sleep_date, start_datetime, levels, duration_in_min, and is_main_sleep.}
}
\description{
Prepares raw data for later processing
}
