% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_as_typical_sleep.R
\name{as_typical_sleep}
\alias{as_typical_sleep}
\title{Typical sleep period}
\usage{
as_typical_sleep(sleep_data)
}
\arguments{
\item{sleep_data}{sleep-levels dataset containing the following columns: person_id, sleep_date, start_datetime, level, duration_in_min, and is_main_sleep.
This is the default schema from the All of Us sleep_levels table.}
}
\value{
a typical sleep object
}
\description{
The typical sleep algorithm returns a typical sleep object that contains the original data plus inserted awake levels and additioanl columns.
The additional columns are the result of the typical sleep algorithm: typical_sleep_date, is_typical_sleep, median_msp, median_sleep_start, median_sleep_end
}
