% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center the time over midnight}
\usage{
center(x)
}
\arguments{
\item{x}{integer on the interval [0,1439]}
}
\description{
Centers on midnight such that Noon to 11:59 has negative sign. \cr
Midnight to Noon as positive sign. \cr
Minutes are mapped thusly: \cr
720:1439 -> -720:-1 \cr
0:720 -> 0:720 \cr
}
