% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sleep_metrics.R
\name{compute_sleep_metrics}
\alias{compute_sleep_metrics}
\title{Compute sleep metrics. All metrics are computed for each person_id and date.}
\usage{
compute_sleep_metrics(sleep_data, date_col)
}
\arguments{
\item{sleep_data}{sleep-levels dataset containing the following columns: person_id, date, start_time, level, duration, and is_main_sleep.}

\item{date_col}{name of date column}
}
\value{
A dataframe with the following columns:
\describe{
  \item{sleep_onset}{The start_datetime of the first sleep segment, where sleep segment levels are not wake, awake, or restless.}
  \item{sleep_offset}{The end datetime of the last sleep segment, where sleep segment levels are not wake, awake, or restless. The end datetime is computed by adding sleep duration to start_datetime.}
  \item{sleep_duration}{Duration of sleep in minutes. eqn{(sleep_offset - sleep_onset) / 60}}
  \item{midsleep_point}{Midpoint between sleep_onset and sleep_offset. eqn{(sleep_onset + sleep_offset) / 2}}
  \item{total_sleep_time}{Sum of all sleep segment durations, where level is not awake, wake, or restless.}
  \item{rem_duration}{Sum of all sleep segment durations, where level is rem.}
  \item{deep_duration}{Sum of all sleep segment durations, where level is deep.}
  \item{light_duration}{Sum of all sleep segment durations, where level is light.}
  \item{pct_rem}{Percentage of rem sleep duration. Denominator is the sum of all sleep segment durations whose level is not awake, wake, or restless. If no rem levels exist, then pct_rem is NA.}
  \item{pct_deep}{Percentage of deep sleep duration. Denominator is the sum of all sleep segment durations whose level is not awake, wake, or restless. If no deep levels exist, then pct_deep is NA.}
  \item{pct_light}{Percentage of light sleep duration. Denominator is the sum of all sleep segment durations whose level is not awake, wake, or restless. If no light levels exist, then pct_light is NA.}
  \item{pct_asleep}{Percentage of asleep sleep levels. Denominator is the sum of all sleep segment durations whose level is not awake, wake, or restless. If no asleep levels exist, then pct_asleep is NA.}
  \item{awake_duration}{Total duration of all sleep segments whose level is awake.}
  \item{wake_duration}{Total duration of all sleep segments whose level is wake.}
  \item{restless_duration}{Total duration of all sleep segments whose level is restless.}
  \item{pct_restless}{Percentage of restless sleep levels. Denominator is the sum of all sleep segment durations.}
  \item{pct_awake}{Percentage of awake sleep levels. Denominator is the sum of all sleep segment durations.}
  \item{pct_wake}{Percentage of wake sleep levels. Denominator is the sum of all sleep segment durations.}
  \item{bedtime}{start_datetime of first sleep segment.}
  \item{waketime}{The end datetime of the final sleep segment.}
  \item{time_in_bed}{Time in bed in minutes. eqn{(bedtime - waketime) / 60}.}
  \item{num_awakenings}{Number of contiguous sleep segments indcating an awakening. Segments of differing levels will be combined to form a single contiguous sleep segment given the level is one of awake, wake, or restless.}
  \item{num_long_awakenings}{Number of wake levels >= 30 minutes.}
  \item{longest_wake_duration}{Longest wake duration in minutes.}
  \item{wake_after_sleep_onset}{Duration in minutes of contiguous segments of awake, wake, and/or restless following at least one segment of sleep.}
  \item{wake_to_end_of_log_latency}{Duration in minutes}
}
}
\description{
Compute sleep metrics. All metrics are computed for each person_id and date.
}
