% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typical.sleep.R
\name{.tsp}
\alias{.tsp}
\title{Typical sleep algorithm}
\usage{
.tsp(all_sleep_dat)
}
\arguments{
\item{all_sleep_dat}{sleep-levels dataset containing the following columns: person_id, sleep_date, start_datetime, level, duration_in_min, and is_main_sleep.}
}
\description{
Typical sleep algorithm
}
