% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\name{example_dataset}
\alias{example_dataset}
\title{Location of the example dataset}
\usage{
example_dataset()
}
\description{
Location of the example dataset
}
\examples{
\dontrun{
# If parsing from JSON format
file_loc <- example_dataset()
out <- parse_fitbit_json(file_loc)
}
}
