% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_relevant_sleep.R
\name{find_relevant_sleep}
\alias{find_relevant_sleep}
\title{Find relevant sleep logs}
\usage{
find_relevant_sleep(all_sleep_dat)
}
\arguments{
\item{all_sleep_dat}{sleep-levels dataset containing the following columns: person_id, date, start_time, level, duration, and is_main_sleep.}
}
\description{
finds relevant sleep logs by computing the median sleep point when is_main_sleep is TRUE and filtering out sleep logs that are outside of that interval.
}
