% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_get_algorithm_data.R
\name{get_algorithm_data}
\alias{get_algorithm_data}
\title{Compute nap metrics for each person_id and date.}
\usage{
get_algorithm_data(sleep_data)
}
\arguments{
\item{sleep_data}{an object returned from as_sleep_logs or typical_sleep}
}
\value{
A dataframe with the following columns (by person_id and date):
\describe{ 
 \item{nap_count}{Number of sleep logs outside of primary sleep period}
 \item{nap_length}{Total duration of sleep across sleep logs outside of the primary sleep period}
}
}
\description{
Compute nap metrics for each person_id and date.
}
\examples{
\dontrun{
# If parsing from JSON format
dat <- parse_fitbit_json("sleep_data.json")
dat_logs <- as_sleep_logs(dat)
tsp <- get_typical_sleep(dat_logs)
get_algorithm_data(tsp)
}

}
