% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sleep_metrics.R
\name{get_naps}
\alias{get_naps}
\title{Function for determining nap counts and duration. Not exported.}
\usage{
get_naps(all_sleep_dat, date_col)
}
\arguments{
\item{all_sleep_dat}{Dataset containing sleep segments. Must contain person_id, is_main_sleep, date_col, level, and start_datetime.}

\item{date_col}{name of date column}
}
\value{
dataframe with person_id, date_col, nap_count, and nap_length
}
\description{
Function for determining nap counts and duration. Not exported.
}
