% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fitbit_json.R
\name{parse_fitbit_json}
\alias{parse_fitbit_json}
\title{Parses the JSON object returned from Google Takeout}
\usage{
parse_fitbit_json(input_file, person_id = 1)
}
\arguments{
\item{input_file}{the Fitbit JSON object}

\item{person_id}{optional person ID. Default is 1.}
}
\description{
Parses the JSON object returned from Google Takeout
}
