% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typical_sleep.R
\name{typical_sleep}
\alias{typical_sleep}
\title{Typical sleep period}
\usage{
typical_sleep(sleep_data)
}
\arguments{
\item{sleep_data}{sleep-levels dataset containing the following columns: person_id, sleep_date, start_datetime, level, duration_in_min, and is_main_sleep.
This is the default schema from the All of Us sleep_levels table.}
}
\value{
Returns the original data with two appended columns, sleep_date_new and tsp. sleep_date_new is the recomputed date of sleep based on the typical sleep period. \cr
tsp is TRUE if the sleep log falls within the typical sleep period. Otherwise, it is false.
}
\description{
The typical sleep period is computed in three steps: 
   1. Finds relevant sleep logs are those that fall within +/-8 hours of the midsleep point. 
   2. Determines the median bedtime and waketime from the relevant logs.
   3. Labels each sleep log as either part of the typical sleep period or not by filtering out those
   that are outside of the bedtime/waketime interval.
}
