% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_get_algorithm_data.R
\name{get_algorithm_data}
\alias{get_algorithm_data}
\title{Compute algorithm metrics}
\usage{
get_algorithm_data(sleep_data)
}
\arguments{
\item{sleep_data}{an object returned from as_sleep_logs or typical_sleep}
}
\value{
A dataframe with the following columns:
\describe{ 
 \item{person_id}{person ID number}
 \item{median_msp}{Median of the median sleep period}
 \item{median_sleep_start}{Median sleep start}
 \item{median_sleep_end}{Median sleep end}
 \item{bt_wt_by_msp}{Indicates whether the bedtime and waketime were determined by MSP +/- 3.5 hours. This happens when the median BT > median WT}
 \item{bedtime_gt_waketime}{Indicates whether the median BT > median WT} 
}
}
\description{
Compute algorithm metrics
}
